+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added the CHANGELOG.md file and CI check for it
+- REMIX-2734: Unselect all objects with ESC
 
 ### Changed