+++ source/extensions/lightspeed.event.capture_persp_to_persp/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.2] - 2024-03-05
+### Changed
+- Changed the merge config reference for the mdl path for tests
+- Now the `.kit` within `trex.app.resources` is referenced
+
 ## [1.1.1] - 2023-12-11
 ### Changed
 - Minor change for an API rename