+++ source/extensions/lightspeed.trex.hotkeys/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.0] - 2024-03-04
+### Added
+- ESC hotkey to unselect all selected objects
+
 ## [0.1.0] - 2024-02-05
 ### Added
 - Created