+++ source/extensions/lightspeed.trex.layout.texturecraft/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.0.2] - 2024-03-05
+### Changed
+- Changed the merge config reference for the mdl path for tests
+- Now the `.kit` within `trex.app.resources` is referenced
 
 ## [2.0.1] - 2023-12-07
 ### Changed