+++ source/extensions/lightspeed.trex.packaging.core/docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.3] - 2024-03-05
+### Changed
+- Changed the merge config reference for the mdl path for tests
+- Now the `.kit` within `trex.app.resources` is referenced
+
 ## [1.0.2] - 2023-10-18
 ### Fixed
 - Fixed crash when opening stage